Rename DraftMouseHandler.tsx to DraftPointerHandler.tsx
